Clean up code a bit

+++ app/models/concerns/distanceable.rb
@@ -5,7 +5,6 @@ module Distanceable
 
   module ClassMethods
     def total_distance(points = nil, unit = :km)
-      # Handle method being called directly on relation vs with array
       if points.nil?
         calculate_distance_for_relation(unit)
       else
@@ -50,20 +49,17 @@ module Distanceable
 
       return 0 if points.length < 2
 
-      # OPTIMIZED: Single SQL query instead of N individual queries
       total_meters = calculate_batch_distances(points).sum
 
       total_meters.to_f / ::DISTANCE_UNITS[unit.to_sym]
     end
 
-    # Optimized batch distance calculation using single SQL query
     def calculate_batch_distances(points)
       return [] if points.length < 2
 
       point_pairs = points.each_cons(2).to_a
       return [] if point_pairs.empty?
 
-      # Create a VALUES clause with all point pairs
       values_clause = point_pairs.map.with_index do |(p1, p2), index|
         "(#{index}, ST_GeomFromEWKT('#{p1.lonlat}')::geography, ST_GeomFromEWKT('#{p2.lonlat}')::geography)"
       end.join(', ')
@@ -71,13 +67,13 @@ module Distanceable
       # Single query to calculate all distances
       results = connection.execute(<<-SQL.squish)
         WITH point_pairs AS (
-          SELECT 
+          SELECT
             pair_id,
             point1,
             point2
           FROM (VALUES #{values_clause}) AS t(pair_id, point1, point2)
         )
-        SELECT 
+        SELECT
           pair_id,
           ST_Distance(point1, point2) as distance_meters
         FROM point_pairs
@@ -85,7 +81,7 @@ module Distanceable
       SQL
 
       # Return array of distances in meters
-      results.map { |row| row['distance_meters'].to_f }
+      results.map { |row| row['distance_meters'].to_i }
     end
   end
 
@@ -94,7 +90,6 @@ module Distanceable
       raise ArgumentError, "Invalid unit. Supported units are: #{::DISTANCE_UNITS.keys.join(', ')}"
     end
 
-    # Extract coordinates based on what type other_point is
     other_lonlat = extract_point(other_point)
     return nil if other_lonlat.nil?
